centralizing and fleshing out tests for phantom selections including backward compatibility
