+++ pkg/phantoms/test/phantom_subnets.toml
@@ -1,0 +1,25 @@
+
+
+[Networks]
+	[Networks.1]
+        Generation = 1
+        [[Networks.1.WeightedSubnets]]
+            Weight = 9
+            Subnets = ["192.122.190.0/24", "2001:48a8:687f:1::/64"] 
+
+    [Networks.2]
+        Generation = 2
+        [[Networks.2.WeightedSubnets]]
+            Weight = 1
+            Subnets = ["192.122.190.0/28", "2001:48a8:687f:1::/96"] 
+
+    [Networks.957]
+        Generation = 957
+        [[Networks.957.WeightedSubnets]]
+            Weight = 9
+            RandomizeDstPort = true
+            Subnets = ["192.122.190.0/24", "2001:48a8:687f:1::/64"] 
+        [[Networks.957.WeightedSubnets]]
+            Weight = 1
+            RandomizeDstPort = false
+            Subnets = ["141.219.0.0/16", "35.8.0.0/16"]
