--- pkg/station/lib/phantom_selector.go
@@ -1,616 +1,0 @@
-package lib
-
-import (
-	"crypto/rand"
-	"crypto/sha256"
-	"encoding/binary"
-	"errors"
-	"fmt"
-	"math/big"
-	mrand "math/rand"
-	"net"
-	"sort"
-	"time"
-
-	wr "github.com/mroth/weightedrand"
-	"golang.org/x/crypto/hkdf"
-)
-
-const (
-	phantomSelectionMinGeneration uint = 1
-	phantomHkdfMinVersion         uint = 2
-)
-
-var (
-	// ErrLegacyAddrSelectBug indicates that we have hit a corner case in a legacy address selection
-	// algorithm that causes phantom address selection to fail.
-	ErrLegacyAddrSelectBug = errors.New("no valid addresses specified")
-	// ErrMissingAddrs indicates that no subnets were provided with addresses to select from. This
-	// is only valid for phantomHkdfMinVersion and newer.
-	ErrMissingAddrs = errors.New("no valid addresses specified to select")
-)
-
-// getSubnetsVarint - return EITHER all subnet strings as one composite array if
-// we are selecting unweighted, or return the array associated with the (seed)
-// selected array of subnet strings based on the associated weights
-//
-// Used by Client version 0 and 1
-func (sc *SubnetConfig) getSubnetsVarint(seed []byte, weighted bool) ([]*phantomNet, error) {
-
-	if weighted {
-		// seed random with hkdf derived seed provided by client
-		seedInt, n := binary.Varint(seed)
-		if n == 0 {
-			return nil, fmt.Errorf("failed to seed random for weighted rand")
-		}
-
-		// nolint:staticcheck // here for backwards compatibility with clients
-		mrand.Seed(seedInt)
-
-		choices := make([]wr.Choice, 0, len(sc.WeightedSubnets))
-		for _, cjSubnet := range sc.WeightedSubnets {
-			cjSubnet := cjSubnet // copy loop ptr
-			choices = append(choices, wr.Choice{Item: &cjSubnet, Weight: uint(cjSubnet.Weight)})
-		}
-		c, err := wr.NewChooser(choices...)
-		if err != nil {
-			return nil, err
-		}
-
-		return parseSubnets(c.Pick().(*ConjurePhantomSubnet))
-
-	}
-
-	// Use unweighted config for subnets, concat all into one array and return.
-	out := []*phantomNet{}
-	for _, cjSubnet := range sc.WeightedSubnets {
-		nets, err := parseSubnets(&cjSubnet)
-		if err != nil {
-			return nil, fmt.Errorf("error parsing subnet: %v", err)
-		}
-		out = append(out, nets...)
-	}
-
-	return out, nil
-}
-
-// getSubnetsHkdf returns EITHER all subnet strings as one composite array if
-// we are selecting unweighted, or return the array associated with the (seed)
-// selected array of subnet strings based on the associated weights. Random
-// values are seeded using an hkdf function to prevent biases introduced by
-// math/rand and varint.
-//
-// Used by Client version 2+
-func (sc *SubnetConfig) getSubnetsHkdf(seed []byte, weighted bool) ([]*phantomNet, error) {
-
-	weightedSubnets := sc.WeightedSubnets
-	if weightedSubnets == nil {
-		return []*phantomNet{}, nil
-	}
-
-	if weighted {
-		choices := make([]*ConjurePhantomSubnet, 0, len(weightedSubnets))
-
-		totWeight := int64(0)
-		for _, cjSubnet := range weightedSubnets {
-			cjSubnet := cjSubnet // copy loop ptr
-			weight := cjSubnet.Weight
-			subnets := cjSubnet.Subnets
-			if subnets == nil {
-				continue
-			}
-
-			totWeight += int64(weight)
-			choices = append(choices, &cjSubnet)
-		}
-
-		// Sort choices assending
-		sort.Slice(choices, func(i, j int) bool {
-			return choices[i].Weight < choices[j].Weight
-		})
-
-		// Naive method: get random int, subtract from weights until you are < 0
-		hkdfReader := hkdf.New(sha256.New, seed, nil, []byte("phantom-select-subnet"))
-		totWeightBig := big.NewInt(totWeight)
-		rndBig, err := rand.Int(hkdfReader, totWeightBig)
-		if err != nil {
-			return nil, err
-		}
-
-		// Decrement rnd by each weight until it's < 0
-		rnd := rndBig.Int64()
-		for _, choice := range choices {
-			rnd -= int64(choice.Weight)
-			if rnd < 0 {
-				return parseSubnets(choice)
-			}
-		}
-
-	}
-
-	// Use unweighted config for subnets, concat all into one array and return.
-	out := []*phantomNet{}
-	for _, cjSubnet := range weightedSubnets {
-		nets, err := parseSubnets(&cjSubnet)
-		if err != nil {
-			return nil, fmt.Errorf("error parsing subnet: %v", err)
-		}
-		out = append(out, nets...)
-	}
-
-	return out, nil
-}
-
-// SubnetFilter - Filter IP subnets based on whatever to prevent specific
-// subnets from inclusion in choice. See v4Only and v6Only for reference.
-type SubnetFilter func([]*net.IPNet) ([]*net.IPNet, error)
-
-// V4Only - a functor for transforming the subnet list to only include IPv4 subnets
-func V4Only(obj []*phantomNet) ([]*phantomNet, error) {
-	out := []*phantomNet{}
-
-	for _, _net := range obj {
-		if ipv4net := _net.IP.To4(); ipv4net != nil {
-			out = append(out, _net)
-		}
-	}
-	return out, nil
-}
-
-// V6Only - a functor for transforming the subnet list to only include IPv6 subnets
-func V6Only(obj []*phantomNet) ([]*phantomNet, error) {
-	out := []*phantomNet{}
-
-	for _, _net := range obj {
-		if _net.IP == nil {
-			continue
-		}
-		if net := _net.IP.To4(); net != nil {
-			continue
-		}
-		out = append(out, _net)
-	}
-	return out, nil
-}
-
-func parseSubnets(phantomSubnets *ConjurePhantomSubnet) ([]*phantomNet, error) {
-	subnets := []*phantomNet{}
-
-	if len(phantomSubnets.Subnets) == 0 {
-		return nil, fmt.Errorf("parseSubnets - no subnets provided")
-	}
-
-	for _, strNet := range phantomSubnets.Subnets {
-		_, parsedNet, err := net.ParseCIDR(strNet)
-		if err != nil {
-			return nil, err
-		}
-		if parsedNet == nil {
-			return nil, fmt.Errorf("failed to parse %v as subnet", parsedNet)
-		}
-
-		subnets = append(subnets, &phantomNet{IPNet: parsedNet, supportRandomPort: phantomSubnets.RandomizeDstPort})
-	}
-
-	return subnets, nil
-}
-
-// NewPhantomIPSelector - create object currently populated with a static map of
-// generation number to SubnetConfig, but this may be loaded dynamically in the
-// future.
-func NewPhantomIPSelector() (*PhantomIPSelector, error) {
-	return GetPhantomSubnetSelector()
-}
-
-// PhantomIP provides a wrapper around net.IP that can be used as a net.IP, while also indicating
-// whether or not the subnet from which the address was selected supports port randomization.
-type PhantomIP struct {
-	*net.IP
-	SupportsPortRand bool
-}
-
-type phantomNet struct {
-	*net.IPNet
-	supportRandomPort bool
-}
-
-func subnetsByVersion(seed []byte, clientLibVer uint, genConfig *SubnetConfig) ([]*phantomNet, error) {
-
-	if clientLibVer < phantomHkdfMinVersion {
-		// Version 0 or 1
-		return genConfig.getSubnetsVarint(seed, true)
-	} else {
-		// Version 2
-		return genConfig.getSubnetsHkdf(seed, true)
-	}
-
-}
-
-// Select - select an ip address from the list of subnets associated with the specified generation
-func (p *PhantomIPSelector) Select(seed []byte, generation uint, clientLibVer uint, v6Support bool) (*PhantomIP, error) {
-	genConfig := p.GetSubnetsByGeneration(generation)
-	if genConfig == nil {
-		return nil, fmt.Errorf("generation number not recognized")
-	}
-
-	genSubnets, err := subnetsByVersion(seed, clientLibVer, genConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	if !v6Support {
-		genSubnets, err = V4Only(genSubnets)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	// handle legacy clientLibVersions for selecting phantoms.
-	if clientLibVer < phantomSelectionMinGeneration {
-		// Version 0
-		ip, err := selectPhantomImplV0(seed, genSubnets)
-		if err != nil {
-			return nil, err
-		}
-		return ip, nil
-	} else if clientLibVer < phantomHkdfMinVersion {
-		// Version 1
-		ip, err := selectPhantomImplVarint(seed, genSubnets)
-		if err != nil {
-			return nil, err
-		}
-		return ip, nil
-	}
-
-	// Version 2+
-	ip, err := selectPhantomImplHkdf(seed, genSubnets)
-	if err != nil {
-		return nil, err
-	}
-	return ip, nil
-}
-
-// selectPhantomImplVarint - select an ip address from the list of subnets
-// associated with the specified generation by constructing a set of start and
-// end values for the high and low values in each allocation. The random number
-// is then bound between the global min and max of that set. This ensures that
-// addresses are chosen based on the number of addresses in the subnet.
-func selectPhantomImplVarint(seed []byte, subnets []*phantomNet) (*PhantomIP, error) {
-	type idNet struct {
-		min, max big.Int
-		net      *phantomNet
-	}
-	var idNets []idNet
-
-	// Compose a list of ID Nets with min, max and network associated and count
-	// the total number of available addresses.
-	addressTotal := big.NewInt(0)
-	for _, _net := range subnets {
-		netMaskOnes, _ := _net.Mask.Size()
-		if ipv4net := _net.IP.To4(); ipv4net != nil {
-			_idNet := idNet{}
-			_idNet.min.Set(addressTotal)
-			addressTotal.Add(addressTotal, big.NewInt(2).Exp(big.NewInt(2), big.NewInt(int64(32-netMaskOnes)), nil))
-			_idNet.max.Sub(addressTotal, big.NewInt(1))
-			_idNet.net = _net
-			idNets = append(idNets, _idNet)
-		} else if ipv6net := _net.IP.To16(); ipv6net != nil {
-			_idNet := idNet{}
-			_idNet.min.Set(addressTotal)
-			addressTotal.Add(addressTotal, big.NewInt(2).Exp(big.NewInt(2), big.NewInt(int64(128-netMaskOnes)), nil))
-			_idNet.max.Sub(addressTotal, big.NewInt(1))
-			_idNet.net = _net
-			idNets = append(idNets, _idNet)
-		} else {
-			return nil, fmt.Errorf("failed to parse %v", _net)
-		}
-	}
-
-	// If the total number of addresses is 0 something has gone wrong
-	if addressTotal.Cmp(big.NewInt(0)) <= 0 {
-		return nil, ErrLegacyAddrSelectBug
-	}
-
-	// Pick a value using the seed in the range of between 0 and the total
-	// number of addresses.
-	id := &big.Int{}
-	id.SetBytes(seed)
-	if id.Cmp(addressTotal) >= 0 {
-		id.Mod(id, addressTotal)
-	}
-
-	// Find the network (ID net) that contains our random value and select a
-	// random address from that subnet.
-	// min >= id%total >= max
-	var result *PhantomIP
-	for _, _idNet := range idNets {
-		// fmt.Printf("tot:%s, seed%%tot:%s     id cmp max: %d,  id cmp min: %d %s\n", addressTotal.String(), id, _idNet.max.Cmp(id), _idNet.min.Cmp(id), _idNet.net.String())
-		if _idNet.max.Cmp(id) >= 0 && _idNet.min.Cmp(id) <= 0 {
-			res, err := SelectAddrFromSubnet(seed, _idNet.net.IPNet)
-			if err != nil {
-				return nil, fmt.Errorf("failed to chose IP address: %v", err)
-			}
-
-			result = &PhantomIP{IP: &res, SupportsPortRand: _idNet.net.supportRandomPort}
-		}
-	}
-
-	// We want to make it so this CANNOT happen
-	if result == nil {
-		return nil, errors.New("nil result should not be possible")
-	}
-	return result, nil
-}
-
-// selectPhantomImplV0 implements support for the legacy (buggy) client phantom
-// address selection algorithm.
-func selectPhantomImplV0(seed []byte, subnets []*phantomNet) (*PhantomIP, error) {
-
-	addressTotal := big.NewInt(0)
-
-	type idNet struct {
-		min, max big.Int
-		net      *phantomNet
-	}
-	var idNets []idNet
-
-	for _, _net := range subnets {
-		netMaskOnes, _ := _net.Mask.Size()
-		if ipv4net := _net.IP.To4(); ipv4net != nil {
-			_idNet := idNet{}
-			_idNet.min.Set(addressTotal)
-			addressTotal.Add(addressTotal, big.NewInt(2).Exp(big.NewInt(2), big.NewInt(int64(32-netMaskOnes)), nil))
-			addressTotal.Sub(addressTotal, big.NewInt(1))
-			_idNet.max.Set(addressTotal)
-			_idNet.net = _net
-			idNets = append(idNets, _idNet)
-		} else if ipv6net := _net.IP.To16(); ipv6net != nil {
-			_idNet := idNet{}
-			_idNet.min.Set(addressTotal)
-			addressTotal.Add(addressTotal, big.NewInt(2).Exp(big.NewInt(2), big.NewInt(int64(128-netMaskOnes)), nil))
-			addressTotal.Sub(addressTotal, big.NewInt(1))
-			_idNet.max.Set(addressTotal)
-			_idNet.net = _net
-			idNets = append(idNets, _idNet)
-		} else {
-			return nil, fmt.Errorf("failed to parse %v", _net)
-		}
-	}
-
-	if addressTotal.Cmp(big.NewInt(0)) <= 0 {
-		return nil, ErrLegacyAddrSelectBug
-	}
-
-	id := &big.Int{}
-	id.SetBytes(seed)
-	if id.Cmp(addressTotal) > 0 {
-		id.Mod(id, addressTotal)
-	}
-
-	var result *PhantomIP
-	for _, _idNet := range idNets {
-		if _idNet.max.Cmp(id) >= 0 && _idNet.min.Cmp(id) == -1 {
-			res, err := SelectAddrFromSubnet(seed, _idNet.net.IPNet)
-			if err != nil {
-				return nil, fmt.Errorf("failed to chose IP address: %v", err)
-			}
-			result = &PhantomIP{IP: &res, SupportsPortRand: _idNet.net.supportRandomPort}
-		}
-	}
-	if result == nil {
-		return nil, ErrLegacyAddrSelectBug
-	}
-	return result, nil
-}
-
-// SelectAddrFromSubnet - given a seed and a CIDR block choose an address.
-//
-// This is done by generating a seeded random bytes up to teh length of the full
-// address then using the net mask to zero out any bytes that are already
-// specified by the CIDR block. Tde masked random value is then added to the
-// cidr block base giving the final randomly selected address.
-func SelectAddrFromSubnet(seed []byte, net1 *net.IPNet) (net.IP, error) {
-	bits, addrLen := net1.Mask.Size()
-
-	ipBigInt := &big.Int{}
-	if v4net := net1.IP.To4(); v4net != nil {
-		ipBigInt.SetBytes(net1.IP.To4())
-	} else if v6net := net1.IP.To16(); v6net != nil {
-		ipBigInt.SetBytes(net1.IP.To16())
-	}
-
-	seedInt, n := binary.Varint(seed)
-	if n == 0 {
-		return nil, fmt.Errorf("failed to create seed ")
-	}
-
-	// nolint:staticcheck // here for backwards compatibility with clients
-	mrand.Seed(seedInt)
-	randBytes := make([]byte, addrLen/8)
-
-	// nolint:staticcheck // here for backwards compatibility with clients
-	_, err := mrand.Read(randBytes)
-	if err != nil {
-		return nil, err
-	}
-	randBigInt := &big.Int{}
-	randBigInt.SetBytes(randBytes)
-
-	mask := make([]byte, addrLen/8)
-	for i := 0; i < addrLen/8; i++ {
-		mask[i] = 0xff
-	}
-	maskBigInt := &big.Int{}
-	maskBigInt.SetBytes(mask)
-	maskBigInt.Rsh(maskBigInt, uint(bits))
-
-	randBigInt.And(randBigInt, maskBigInt)
-	ipBigInt.Add(ipBigInt, randBigInt)
-
-	return net.IP(ipBigInt.Bytes()), nil
-}
-
-// SelectAddrFromSubnetOffset given a CIDR block and offset, return the net.IP
-//
-// Version 2: HKDF-based
-func SelectAddrFromSubnetOffset(net1 *net.IPNet, offset *big.Int) (net.IP, error) {
-	bits, addrLen := net1.Mask.Size()
-
-	// Compute network size (e.g. an ipv4 /24 is 2^(32-24)
-	var netSize big.Int
-	netSize.Exp(big.NewInt(2), big.NewInt(int64(addrLen-bits)), nil)
-
-	// Check that offset is within this subnet
-	if netSize.Cmp(offset) <= 0 {
-		return nil, errors.New("Offset too big for subnet")
-	}
-
-	ipBigInt := &big.Int{}
-	if v4net := net1.IP.To4(); v4net != nil {
-		ipBigInt.SetBytes(net1.IP.To4())
-	} else if v6net := net1.IP.To16(); v6net != nil {
-		ipBigInt.SetBytes(net1.IP.To16())
-	}
-
-	ipBigInt.Add(ipBigInt, offset)
-
-	return net.IP(ipBigInt.Bytes()), nil
-}
-
-// selectPhantomImplHkdf selects an ip address from the list of subnets
-// associated with the specified generation by constructing a set of start and
-// end values for the high and low values in each allocation. The random number
-// is then bound between the global min and max of that set. This ensures that
-// addresses are chosen based on the number of addresses in the subnet.
-func selectPhantomImplHkdf(seed []byte, subnets []*phantomNet) (*PhantomIP, error) {
-	type idNet struct {
-		min, max big.Int
-		net      *phantomNet
-	}
-	var idNets []idNet
-
-	// Compose a list of ID Nets with min, max and network associated and count
-	// the total number of available addresses.
-	addressTotal := big.NewInt(0)
-	for _, _net := range subnets {
-		netMaskOnes, _ := _net.Mask.Size()
-		if ipv4net := _net.IP.To4(); ipv4net != nil {
-			_idNet := idNet{}
-			_idNet.min.Set(addressTotal)
-			addressTotal.Add(addressTotal, big.NewInt(2).Exp(big.NewInt(2), big.NewInt(int64(32-netMaskOnes)), nil))
-			_idNet.max.Sub(addressTotal, big.NewInt(1))
-			_idNet.net = _net
-			idNets = append(idNets, _idNet)
-		} else if ipv6net := _net.IP.To16(); ipv6net != nil {
-			_idNet := idNet{}
-			_idNet.min.Set(addressTotal)
-			addressTotal.Add(addressTotal, big.NewInt(2).Exp(big.NewInt(2), big.NewInt(int64(128-netMaskOnes)), nil))
-			_idNet.max.Sub(addressTotal, big.NewInt(1))
-			_idNet.net = _net
-			idNets = append(idNets, _idNet)
-		} else {
-			return nil, fmt.Errorf("failed to parse %v", _net)
-		}
-	}
-
-	// If the total number of addresses is 0 something has gone wrong
-	if addressTotal.Cmp(big.NewInt(0)) <= 0 {
-		return nil, ErrMissingAddrs
-	}
-
-	// Pick a value using the seed in the range of between 0 and the total
-	// number of addresses.
-	hkdfReader := hkdf.New(sha256.New, seed, nil, []byte("phantom-addr-id"))
-	id, err := rand.Int(hkdfReader, addressTotal)
-	if err != nil {
-		return nil, err
-	}
-
-	// Find the network (ID net) that contains our random value and select a
-	// random address from that subnet.
-	// min >= id%total >= max
-	var result *PhantomIP
-	for _, _idNet := range idNets {
-		// fmt.Printf("tot:%s, seed%%tot:%s     id cmp max: %d,  id cmp min: %d %s\n", addressTotal.String(), id, _idNet.max.Cmp(id), _idNet.min.Cmp(id), _idNet.net.String())
-		if _idNet.max.Cmp(id) >= 0 && _idNet.min.Cmp(id) <= 0 {
-
-			var offset big.Int
-			offset.Sub(id, &_idNet.min)
-			res, err := SelectAddrFromSubnetOffset(_idNet.net.IPNet, &offset)
-			if err != nil {
-				return nil, fmt.Errorf("failed to chose IP address: %v", err)
-			}
-
-			result = &PhantomIP{IP: &res, SupportsPortRand: _idNet.net.supportRandomPort}
-		}
-	}
-
-	// We want to make it so this CANNOT happen
-	if result == nil {
-		return nil, errors.New("nil result should not be possible")
-	}
-	return result, nil
-}
-
-// GetSubnetsByGeneration - provide a generation index. If the generation exists
-// the associated SubnetConfig is returned. If it is not defined the default
-// subnets are returned.
-func (p *PhantomIPSelector) GetSubnetsByGeneration(generation uint) *SubnetConfig {
-	if subnets, ok := p.Networks[generation]; ok {
-		return subnets
-	}
-
-	// No Default subnets provided if the generation is not known
-	return nil
-}
-
-// AddGeneration - add a subnet config as a new new generation, if the requested
-// generation index is taken then it uses (and returns) the next available
-// number.
-func (p *PhantomIPSelector) AddGeneration(gen int, subnets *SubnetConfig) uint {
-
-	ugen := uint(gen)
-
-	if gen == -1 || p.IsTakenGeneration(ugen) {
-		ugen = p.newGenerationIndex()
-	}
-
-	p.Networks[ugen] = subnets
-	return ugen
-}
-
-func (p *PhantomIPSelector) newGenerationIndex() uint {
-	maxGen := uint(0)
-	for k := range p.Networks {
-		if k > maxGen {
-			maxGen = k
-		}
-	}
-	return maxGen + 1
-}
-
-// IsTakenGeneration - check if the generation index is already in use.
-func (p *PhantomIPSelector) IsTakenGeneration(gen uint) bool {
-	if _, ok := p.Networks[gen]; ok {
-		return true
-	}
-	return false
-}
-
-// RemoveGeneration - remove a generation from the mapping
-func (p *PhantomIPSelector) RemoveGeneration(generation uint) bool {
-	p.Networks[generation] = nil
-	return true
-}
-
-// UpdateGeneration - Update the subnet list associated with a specific generation
-func (p *PhantomIPSelector) UpdateGeneration(generation uint, subnets *SubnetConfig) bool {
-	p.Networks[generation] = subnets
-	return true
-}
-
-func init() {
-	// NOTE: math/rand is only used for backwards compatibility.
-	// nolint:staticcheck
-	mrand.Seed(time.Now().UnixNano())
-}
